% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_clean.R
\name{clean_student_data}
\alias{clean_student_data}
\title{Clean student dataset}
\usage{
clean_student_data(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
Cleaned dataframe
}
\description{
Clean student dataset
}
