% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_clean.R
\name{load_student_data}
\alias{load_student_data}
\title{Load a cleaned student dataset}
\usage{
load_student_data()
}
\value{
A cleaned dataframe of student data
}
\description{
Load a cleaned student dataset
}
