% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_relationship}
\alias{plot_relationship}
\title{Plot Relationship Between Any Two Variables}
\usage{
plot_relationship(df, x_var, y_var)
}
\arguments{
\item{df}{A cleaned data frame.}

\item{x_var}{A string representing the name of the variable for the x-axis.}

\item{y_var}{A string representing the name of the variable for the y-axis.}
}
\value{
A ggplot object.
}
\description{
Creates a flexible scatterplot that allows users to
visualize the relationship between any two numeric variables, with a trend line.
}
\examples{
df_gpa <- scale_exam_score(student_data)
plot_relationship(df_gpa, "Attendance", "gpa")
}
