% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{sleep_study_relationship}
\alias{sleep_study_relationship}
\title{Analyze relationship between sleep and study time}
\usage{
sleep_study_relationship(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
Correlation coefficient and interpretation
}
\description{
Returns correlation coefficient between hours of sleep and hours spent studying, based on observations in a dataset.
}
